Commit
Got some more functionality into it.  It now will opens the categories
with better performance for markers you can click on them.  Also added
the ability for someone to specify which view of the map they want it
to open to either map or satelite.  Moved some of the images around to
make them easier to use in the folder structure.