Embed
Added all of the options to embed the map into a page. An example can
be seen at http://www2.byui.edu/testFolder/master-framework/content.html

Working on commenting all of the code and I structured the code to work
better for people looking at it on GitHub.com

Also added the functionality of pulling up a building/place whenever a
code is passed in the url.  I forgot to do that in my last commit.