House Cleaning
There was a lot of stuff in here that wasn't really all that necessary
so I deleted it and just kept the things that were really necessary.  I
also have restructured the folder structure to be easier for people to
navigate through it.  Before it was structured after the Ingeniux
structure but I realized this wasn't the greatest for Github.  And when
someone wants to put up a new version they should be compitant enough
to put everything in the right place.